Update cargo-audit to v0.21.0
Work around a compatibility issue between our previous version and the cr132
rust toolchain due to the Cargo.lock version number being bumped to 4.

This is the contents of the published package downloaded and unpacked from
https://crates.io/api/v1/crates/cargo-audit/0.21.0/download with the
README.chromium updated manually.

We build this in CI with cargo, so updating other vendored deps is not
necessary.