[cr132] `PersistedData` with a getter for `PrefService`
This change is addressing one of the upstream leaky pointers for
`extensions::ChromeUpdateClientConfig`, however it doesn't touch on the
main other updaters for the browser, and just provides a simple getter
that resturns the pref service.

This change copies the simpler approach for now, until the other
dangling pointers are corrected.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/f8c1ffc702ef9a40fcc7aeedf1bff3093cc0a6e3

commit f8c1ffc702ef9a40fcc7aeedf1bff3093cc0a6e3
Author: Joshua Pawlicki <waffles@chromium.org>
Date:   Thu Nov 7 19:44:45 2024 +0000

    Extension Updater: Refetch PrefService from BrowserContext each call.

    This ensures that we don't retain a dangling reference to PrefService.

    Fixed: 374916448, 40274742